// src/interfaces/FilterFieldItem.ts
interface FilterFieldItem {
  FieldName: string
  Value?: string
  Visible?: boolean
  DisplayName?: string
  Filter?: boolean
  Control?: string
  DataType?: string
  DropdownSource?: string
  Selected?: string
  Predicate?: string
  FilterValue?: string
  FilterValue2?: string
  Sort?: string
  Shown?: boolean
  Options?: Array<any>
}
export { FilterFieldItem }